+++ R/hop-index-common.R
@@ -31,16 +31,6 @@ hop_index_common <- function(x,
   args <- vec_cast_common(i, !!!args)
   args <- lapply(args, vec_proxy_compare)
 
-  # Early exit if empty input
-  # (but after all size checks have been done)
-  if (size == 0L) {
-    return(vec_init(ptype, 0L))
-  }
-
-  if (x_size == 0L) {
-    return(vec_init(ptype, size))
-  }
-
   i <- args[[1L]]
   starts <- args[[2L]]
   stops <- args[[3L]]